requirements.txt
alabaster==0.7.12
anyio==3.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.1
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
black==21.12b0
bleach==4.1.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
colorlover==0.3.0
cookiecutter==1.7.3
coverage==6.2
cryptography==36.0.1
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.26
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
docformatter==1.4
docrepr==0.1.1
docutils==0.17.1
entrypoints==0.3
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastiqa==0.6.1
fastprogress==1.0.0
ffmpeg-progress-yield==0.2.0
ffmpeg-python==0.2.0
ffmpeg-quality-metrics==2.2.0
flake8==4.0.1
fonttools==4.28.4
future==0.18.2
h5py==3.6.0
idna==3.3
imageio==2.13.5
imagesize==1.3.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.17.2
jeepney==0.7.1
jellyfish==0.8.9
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.3.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.9.0
jupyterlab-widgets==1.0.2
keyring==23.4.0
kiwisolver==1.3.2
langcodes==3.3.0
lazy-object-proxy==1.7.1
libcst==0.3.23
libsass==0.21.0
line-profiler==3.3.1
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
memory-profiler==0.60.0
meson==0.60.3
mistune==0.8.4
MonkeyType==21.5.0
mpmath==1.2.1
murmurhash==1.0.6
mypy==0.930
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
ninja==1.10.2.3
notebook==6.4.6
numpy==1.21.4
numpydoc==1.1.0
opencv-contrib-python==4.5.4.60
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.7.0
pathspec==0.9.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
plotly==5.4.0
pluggy==1.0.0
poyo==0.5.0
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.10.0
pylint==2.12.2
pyls-black==0.4.7
pyls-spyder==0.4.0
pyparsing==3.0.6
PyQt5==5.12.3
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-html==3.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-black==1.0.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.1.1
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.1
qtconsole==5.2.2
QtPy==2.0.0
qtsass==0.3.0
requests==2.26.0
rope==0.22.0
Rtree==0.9.7
scikit-image==0.19.1
scikit-learn==1.0.1
scipy==1.7.3
SecretStorage==3.3.1
Send2Trash==1.8.0
setupmeta==3.3.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.2.1
spyder-kernels==2.2.0
spyder-line-profiler==0.2.1
spyder-memory-profiler==0.2.1
spyder-unittest==0.4.1
srsly==2.4.2
sureal==0.6.2
sympy==1.9
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.2
thinc==8.0.13
threadpoolctl==3.0.0
three-merge==0.1.1
tifffile==2021.11.2
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.3
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing-extensions==4.0.1
typing-inspect==0.7.1
ujson==5.1.0
untokenize==0.1.1
urllib3==1.26.7
wasabi==0.9.0
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==3.5.2
wrapt==1.13.3
wurlitzer==3.0.2
yapf==0.31.0
zipp==3.6.0
